# src/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RVM API",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Return all resources",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Create resource",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "organisation_id",
                        "in": "query",
                        "description": "Resource organisation id.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Resource name.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "resource_type",
                        "in": "query",
                        "description": "Resource resource_type.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Resource quantity.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "county",
                        "in": "query",
                        "description": "Resource county.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Resource city.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Resource address.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "comments",
                        "in": "query",
                        "description": "Resource comments.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "added_by",
                        "in": "query",
                        "description": "Resource added by.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/resources/{id}": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Show resource info ",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Resources"
                ],
                "summary": "Update resource",
                "operationId": "update",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Resources"
                ],
                "summary": "Delete resource",
                "operationId": "delete",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/resources/list": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "List resources",
                "operationId": "delete",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/resources/categories": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "List resource categories",
                "operationId": "delete",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}